- Making methods static, plot and inventory manager creates instances of plot and inventory - everything else is static.
