--- src/main/java/Action_Classes/PlotManager.java
@@ -1,17 +1,0 @@
-package Action_Classes;
-import Data_Classes.Plot;
-import Data_Classes.Products;
-
-import java.util.ArrayList;
-
-public class PlotManager {
-    private static ArrayList<Plot> myPlots;
-
-    public ArrayList<Plot> getMyPlots() {
-        return myPlots;
-    }
-
-    public void addPlot(){
-        // TO DO
-    }
-}
