+++ src/main/java/Use_Cases/PlotManager.java
@@ -1,0 +1,20 @@
+package Use_Cases;
+import Entities.Plot;
+
+import java.util.ArrayList;
+
+public class PlotManager {
+    private static ArrayList<Plot> myPlots = new ArrayList<>();
+
+    public static ArrayList<Plot> getMyPlots() {
+        return myPlots;
+    }
+
+    public static void addPlot(){
+        // TO DO
+    }
+
+    public static void createNewPlot(){
+        // create a new plot instance and add it to myPLOTS
+    }
+}
